requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
pdfplumber
sentence-transformers
streamlit
python-dotenv
langchain-huggingface>=0.2.0